[UPDATE] Update the room when a shorter path is found

+++ include/amazed.h
@@ -7,14 +7,15 @@
 
 #ifndef AMAZED_H_
     #define AMAZED_H_
+    #include <stddef.h>
 
 typedef struct path_s {
     char *start_room;
     char *end_room;
     char **shortest_path;
     char **current_path;
-    int shortest_path_length;
-    int current_room_number;
+    size_t shortest_path_length;
+    size_t current_room_number;
 } path_t; 
 
 int amazed(void);
